refactor : fix naming convention of imported UI file in app router - remove 'Page' word
